packages/budget-gate/pyproject.toml
[project]
name = "madfam-budget-gate"
version = "0.1.0"
description = "HITL budget gate + pricing-model-agnostic cost estimator for autonomous LLM spenders across the MADFAM ecosystem."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Innovaciones MADFAM" }]

dependencies = [
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/madfam_budget_gate"]

[tool.ruff]
line-length = 100
target-version = "py311"
